Modulize icon-ids
