+++ image-divider.js
@@ -4,6 +4,8 @@ var validateConfData = require('image-divider').utils.validateConfData;
 var _ = require('lodash');
 var pathModule = require('path');
 
+var ICON_DATA_SOURCE = require('lib/icon-ids').ICON_DATA_SOURCE;
+
 
 var SOURCE_IMAGE_PATH = pathModule.join(__dirname, '/materials/images/denzi/Denzi140330-12-1--32x32.png');
 var DEST_ROOT = pathModule.join(__dirname, '/public/dist/icons');
@@ -19,17 +21,11 @@ var ICON_POS_MAP = {
   items: [ICON_HEIGHT * 12, ICON_WIDTH * 16]
 };
 
-
 /**
  * @param {Array} cellIndex [row, col] で最左上を [0, 0] とするアイコン位置
  * @return {object} {image-dividerのimage}
  */
-var _iconIds = [];
 var mapIcon = function mapImage(category, cellIndex, iconId) {
-  if (_iconIds.indexOf(iconId) > -1) {
-    throw new Error(iconId + ' is duplicated icon-id');
-  }
-  _iconIds.push(iconId);
   var basePos = ICON_POS_MAP[category];
   return {
     src: SOURCE_IMAGE_PATH,
@@ -42,53 +38,8 @@ var mapIcon = function mapImage(category, cellIndex, iconId) {
   };
 };
 
-var icons = [
-
-  // characters
-  ['characters', [0, 3], 'alchemist'],
-  ['characters', [0, 4], 'cleric'],
-  ['characters', [0, 0], 'fighter'],
-  ['characters', [1, 1], 'healer'],
-  ['characters', [1, 3], 'knight'],
-  ['characters', [0, 1], 'mage'],
-  ['characters', [0, 8], 'nerd'],
-  ['characters', [0, 2], 'ranger'],
-  ['characters', [0, 5], 'thief'],
-
-  // enemies
-  ['enemies', [3, 13], 'goblin'],
-  ['enemies', [1, 3], 'minotaur'],
-
-  // icons
-  ['icons', [1, 1], 'brain'],
-  ['icons', [1, 2], 'flash'],
-  ['icons', [0, 14], 'happy'],
-  ['icons', [0, 0], 'heart'],
-  ['icons', [4, 15], 'invalid'],
-  ['icons', [3, 0], 'running'],
-  ['icons', [0, 13], 'unhappy'],
-
-  // items
-  ['items', [0, 6], 'arrow'],
-  ['items', [0, 5], 'bow'],
-  ['items', [0, 4], 'gold'],
-  ['items', [1, 1], 'jewel'],
-  ['items', [0, 10], 'jewel_box'],
-  ['items', [2, 14], 'katana'],
-  ['items', [0, 14], 'key'],
-  ['items', [4, 12], 'lantern'],
-  ['items', [4, 11], 'torch'],
-  ['items', [0, 1], 'sword'],
-
-  // objects
-  ['objects', [2, 7], 'chest'],
-  ['objects', [0, 12], 'downstairs'],
-  ['objects', [0, 1], 'signboard'],
-  ['objects', [0, 6], 'treasure_box'],
-  ['objects', [0, 11], 'upstairs']//,
-
-].map(function(imageData) {
-  return mapIcon.apply(null, imageData);
+var icons = ICON_DATA_SOURCE.map(function(iconData) {
+  return mapIcon.apply(null, iconData);
 });
 
 var conf = {
